Time: 0 ms (100.00%) | Memory: 42.5 MB (15.54%) - LeetSync

+++ 2050-count-good-numbers/count-good-numbers.java
@@ -1,6 +1,6 @@
 class Solution {
     long mod = 1000000007;
-    public long pow( int x, long n){
+    public long pow( long x, long n){
         if(n==0)return 1;
 
         long half = pow (x, n/2);
